This is a big commit, most of the changes are for new profile feature, io and lock. io profile support genearl sync/async file io and socket io, include various counters like io complete time, count per io object, r/w counters, max/min latency, it should be enough for general io profiling work. lock profile only support critical section and shared rw lock, since these two types are easy to cause performance issue, they are used for high conention scenario, others like event/mutex are not support currently. the io profile is used to find what io object (file/socket) can cause performance problem. hlep us to find most busy accessed io objects, this give us hint how to optimize their access. the lock profile is used to find what lock are acquired/released in high frequency.
this commit also include some minor change, CPU function/module features are dropped since
they are not very useful.

finally, socket io profiling is not stable for all kinds of application, it can crash.

+++ apsbtr/bitmap.c
@@ -4,6 +4,7 @@
 // Copyright(C) 2009-2010
 //
 
+#include "apsbtr.h"
 #include "bitmap.h"
 #include "util.h"
 #include <assert.h>
@@ -124,7 +125,7 @@ BtrFindFirstClearBitBackward(
 
 	ASSERT(FromIndex < BitMap->SizeOfBitMap);
 
-	for (Index = FromIndex; Index != -1; Index -= 1) {
+	for (Index = FromIndex; Index != (ULONG)-1; Index -= 1) {
 		Status = BtrTestBit(BitMap, Index);
 		if (Status != TRUE) {
 			return Index;
